Adding NuGet and AppVeyor

+++ NuGet/package.ps1
@@ -39,4 +39,7 @@ Foreach ($nuspec in $(Get-Item *.nuspec))
 }
 
 # Delete compiled temporary nuspec.
-Remove-Item $compiledNuspec
+If (Test-Path $compiledNuspec)
+{
+	Remove-Item $compiledNuspec
+}
